migrate kaldi resample tests to functional
